Created loadTeeSheet() and addTeeTime()
loadTeeSheet accepts two String parameters for date
and time where time represents the tee time that
is expanded for viewing the details. To view regular
tee sheet with no expanded tee times, just use empty
string as the time parameter and the method will skip
over the expansion part.

Add tee time accepts a string to specify the time
being edited and a HashMap<String,String> containing
the database column name as the key and entered info
as the value. This method can be used as the edit
and remove function as well.