Updated toString method to fix 'null' value in MySQL
includes addition of method named formatMySqlSyntax which adds
quotations only to non-null values in SingleTeeTime object. The toString
does not have any added quotation marks in the form "\'" as it did
before. (Except for the time variable since that is never null)

Tested minimally but was successful for all tests.