Made the Blocking page functional.
Errors will be thrown if any of the fields are empty, but the program
will continue to run as normal. My plan is to have pop-up messages for
these errors. Also, while we were meeting, we figured out how to write
NULL values into the database. There is a quick example of how to do
this in the SingleTeeTime toString() method.  We also found some issues
with my IDE importing the project from GitHub, so we had to move the
module around. The file is unchanged and should be in the same place
after this commit.